Initial Commit (see description)
This is the initial commit on a new repository. Why did I make a new repository? Because the last one got completely fucked by my own stupidity and not even git revert or reset could save me. That is why the change log all ready has changes logged inside it